System.Drawing: GdipGetStringFormatDigitSubstitution takes a LANGID*
libgdiplus writes the language back through that argument - it null-checks it
and then does a 16-bit store - but the managed declaration called it an int.
Nothing was ever passed there but zero, since StringFormat.language is never
assigned by any path, so on an engine that widens a 4-byte argument into a
clear 8-byte register it read as NULL and libgdiplus skipped the store.

The interpreter does not widen it. A signature of three pointer-sized
parameters takes MINT_ICALL_PPP_P, which reads each argument as sp[i].data.p -
eight bytes out of a slot the int wrote four of - so whatever the slot held
before rides along in the top half. The result is a non-null pointer built from
stack residue, and libgdiplus stores through it.

Pass IntPtr.Zero, which is what the native side reads as "do not report the
language" and is the only thing this code ever wanted.

+++ mcs/class/System.Drawing/System.Drawing/gdipFunctions.cs
@@ -1636,7 +1636,11 @@ namespace System.Drawing
 		[DllImport (GdiPlus)]
 		internal static extern Status GdipSetStringFormatTabStops(IntPtr format, float firstTabOffset, int count, float [] tabStops);
 		[DllImport (GdiPlus)]
-		internal static extern Status GdipGetStringFormatDigitSubstitution(IntPtr format, int language, out StringDigitSubstitute substitute);
+		// language is an out LANGID*, and NULL means "don't report it". Declaring
+		// it as an int passed a value the callee writes through, which only ever
+		// worked because the value was always zero and the upper half of the
+		// register happened to be clear.
+		internal static extern Status GdipGetStringFormatDigitSubstitution(IntPtr format, IntPtr language, out StringDigitSubstitute substitute);
 		[DllImport (GdiPlus)]
 		internal static extern Status GdipSetStringFormatDigitSubstitution(IntPtr format, int language, StringDigitSubstitute substitute);
 		[DllImport (GdiPlus)]
